requirements.txt
bs4>=0.0.2
cdsapi>=0.7.4
ETS_CookBook>=1.2.4
matplotlib>=3.9.2
numpy>=2.1.3
pandas>=2.2.3
python-box>=7.2.0
requests>=2.32.3
tomli_w>=1.2.0
tqdm>=4.67.0